requirements.txt
feedparser>=6.0
anthropic>=0.40
jinja2>=3.1
markdown>=3.4
pyyaml>=6.0
python-dotenv>=1.0
